AlgoStack/requirements.txt
# ═══════════════════════════════════════════════════════════════════════
# AlgoStack v10.2 | Author: Ridhaant Ajoy Thackur
# requirements.txt — Install: pip install -r requirements.txt
# ═══════════════════════════════════════════════════════════════════════

# CORE DATA & TRADING
yfinance>=0.2.40
pandas>=2.1.0
numpy>=1.26.0
pytz>=2024.1
requests>=2.31.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# DASHBOARD & VISUALIZATION
dash>=2.17.0
plotly>=5.22.0
dash-bootstrap-components>=1.6.0

# ENTERPRISE AUTH (email + TOTP 2FA) — set ALGOSTACK_AUTH_ENABLED=1
flask-login>=0.6.3
pyotp>=2.9.0

# IPC & MESSAGING
pyzmq>=26.0.0

# TUNNEL
pyngrok>=7.0.0

# UI & LOGGING
rich>=13.7.0
loguru>=0.7.2

# CONFIG
python-dotenv>=1.0.0

# MARKET CALENDAR
holidays>=0.46

# INTERNET SPEED MONITORING
speedtest-cli>=2.1.3

# CRYPTO PRICE FEED — Binance WebSocket primary (1s latency)
websocket-client>=1.7.0

# COMMODITY PRICE FEED — TradingView WS + fallback scrapers
beautifulsoup4>=4.12.0

# NEWS & SENTIMENT (optional — comment out if not using news dashboard)
feedparser>=6.0.10
vaderSentiment>=3.3.2

# ── PERFORMANCE ACCELERATION ────────────────────────────────────────────────
# Numba JIT: 5-10x faster sweep kernels (CPU, no GPU required)
# Install: pip install numba
numba>=0.59.0

# GPU acceleration via CuPy (requires NVIDIA GPU + CUDA 12)
# GTX 1650 uses CUDA 12 — install ONE of these:
#   pip install cupy-cuda12x        # CUDA 12.x (most common)
#   pip install cupy-cuda11x        # CUDA 11.x (older driver)
# Uncomment the line below if you have CUDA installed:
# cupy-cuda12x>=13.0.0

# ── OPTIONAL ENHANCEMENTS ───────────────────────────────────────────────────
# newsapi-python>=0.2.7    # for richer news feed
# praw>=7.7.0              # Reddit sentiment
# aiohttp>=3.9.0           # async HTTP
